Feat: 문장 조회 시 사용자 조건 추가
