+++ features/phrases/components/templates/PhraseViewPageTemplate.tsx
