+++ features/phrases/services/phrases.service.ts
@@ -1,6 +1,7 @@
 "use server";
 
 import db from "@/lib/db";
+import {getSessionUser} from "@/utils/authUtils";
 
 interface GetPhrasesParams {
   searchType?: string | undefined;
@@ -10,6 +11,11 @@ interface GetPhrasesParams {
 
 // export async function getPhrases(page: number) {
 export async function getPhrases({searchType, keyword, order}: GetPhrasesParams) {
+  const user = await getSessionUser();
+  if (!user) {
+    throw new Error("로그인이 필요합니다.");
+  }
+
   const orderColumn = order === "asc" ? "asc" : "desc";
   const isCorrectSearchType =
     !!searchType && ["japanese", "romaji", "pronunciation", "translation"].includes(searchType);
@@ -34,6 +40,7 @@ export async function getPhrases({searchType, keyword, order}: GetPhrasesParams)
             },
           }
         : {}),
+      userId: user.id,
     },
     // skip: page * 1,
     // take: 1,
@@ -54,3 +61,35 @@ export const deletePhrase = async (id: number) => {
   });
   // } catch (e) {}
 };
+
+interface GetPhraseParams {
+  id: string;
+}
+
+export async function getPhrase({id}: GetPhraseParams) {
+  if (!id || isNaN(Number(id))) {
+    return null;
+  }
+
+  const phrase = await db.phrase.findUnique({
+    where: {
+      id: Number(id),
+    },
+    select: {
+      id: true,
+      japanese: true,
+      romaji: true,
+      pronunciation: true,
+      translation: true,
+      createdAt: true,
+      updatedAt: true,
+      description: true,
+    },
+  });
+
+  if (!phrase) {
+    return null;
+  }
+
+  return phrase;
+}
